Moved all parameters components to Parameters.cs
