+++ src/RhinoInside.Revit.GH/GH/Components/Element/Element.cs
@@ -1,11 +1,6 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Windows.Forms;
-using Grasshopper.GUI;
-using Grasshopper.GUI.Canvas;
 using Grasshopper.Kernel;
-using Grasshopper.Kernel.Attributes;
 using Grasshopper.Kernel.Types;
 using DB = Autodesk.Revit.DB;
 
@@ -192,271 +187,6 @@ namespace RhinoInside.Revit.GH.Components
       DA.SetDataList(0, meshes?.Select((x) =>    new GH_Mesh(x)));
       DA.SetDataList(1, materials?.Select((x) => new GH_Material(x)));
       DA.SetDataList(2, wires?.Select((x) =>     new GH_Curve(x)));
-    }
-  }
-
-  public class ElementParameters : ElementGetter
-  {
-    public override Guid ComponentGuid => new Guid("44515A6B-84EE-4DBD-8241-17EDBE07C5B6");
-    static readonly string PropertyName = "Parameters";
-
-    public ElementParameters() : base(PropertyName) { }
-
-    protected override void RegisterInputParams(GH_InputParamManager manager)
-    {
-      base.RegisterInputParams(manager);
-      manager[manager.AddTextParameter("Name", "N", "Filter params by Name", GH_ParamAccess.item)].Optional = true;
-      manager[manager.AddParameter(new Parameters.Param_Enum<Types.BuiltInParameterGroup>(), "Group", "G", "Filter params by the group they belong", GH_ParamAccess.item)].Optional = true;
-      manager[manager.AddBooleanParameter("ReadOnly", "R", "Filter params by its ReadOnly property", GH_ParamAccess.item)].Optional = true;
-    }
-
-    protected override void RegisterOutputParams(GH_OutputParamManager manager)
-    {
-      manager.AddParameter(new Parameters.ParameterKey(), "Parameters", "P", "Element parameters", GH_ParamAccess.list);
-    }
-
-    protected override void TrySolveInstance(IGH_DataAccess DA)
-    {
-      DB.Element element = null;
-      if (!DA.GetData(ObjectType.Name, ref element))
-        return;
-
-      string parameterName = null;
-      bool noFilterName = (!DA.GetData("Name", ref parameterName) && Params.Input[1].Sources.Count == 0);
-
-      var builtInParameterGroup = DB.BuiltInParameterGroup.INVALID;
-      bool noFilterGroup = (!DA.GetData("Group", ref builtInParameterGroup) && Params.Input[2].Sources.Count == 0);
-
-      bool readOnly = false;
-      bool noFilterReadOnly = (!DA.GetData("ReadOnly", ref readOnly) && Params.Input[3].Sources.Count == 0);
-
-      List<DB.Parameter> parameters = null;
-      if (element is object)
-      {
-        parameters = new List<DB.Parameter>(element.Parameters.Size);
-        foreach (var group in element.GetParameters(RevitAPI.ParameterSet.Any).GroupBy((x) => x.Definition?.ParameterGroup ?? DB.BuiltInParameterGroup.INVALID).OrderBy((x) => x.Key))
-        {
-          foreach (var param in group.OrderBy(x => x.Id.IntegerValue))
-          {
-            if (!noFilterName && parameterName != param.Definition?.Name)
-              continue;
-
-            if (!noFilterGroup && builtInParameterGroup != (param.Definition?.ParameterGroup ?? DB.BuiltInParameterGroup.INVALID))
-              continue;
-
-            if (!noFilterReadOnly && readOnly != param.IsReadOnly)
-              continue;
-
-            parameters.Add(param);
-          }
-        }
-      }
-
-      DA.SetDataList("Parameters", parameters);
-    }
-  }
-
-  public class ElementDecompose : Component, IGH_VariableParameterComponent
-  {
-    public override Guid ComponentGuid => new Guid("FAD33C4B-A7C3-479B-B309-8F5363B25599");
-    public ElementDecompose() : base("Element.Decompose", "Decompose", "Decomposes an Element into its parameters", "Revit", "Element") { }
-
-    public override void AppendAdditionalMenuItems(ToolStripDropDown menu)
-    {
-      base.AppendAdditionalMenuItems(menu);
-
-      bool hasInputData = !Params.Input[0].VolatileData.IsEmpty;
-      bool hasOutputParameters = Params.Output.Count > 0;
-
-      Menu_AppendSeparator(menu);
-      Menu_AppendItem(menu, "Get common parameters", Menu_PopulateOutputsWithCommonParameters, hasInputData, false);
-      Menu_AppendItem(menu, "Get all parameters", Menu_PopulateOutputsWithAllParameters, hasInputData, false);
-      Menu_AppendItem(menu, "Remove unconnected parameters", Menu_RemoveUnconnectedParameters, hasOutputParameters, false);
-    }
-
-    class ComponentAttributes : GH_ComponentAttributes
-    {
-      public ComponentAttributes(IGH_Component component) : base(component) { }
-
-      public override GH_ObjectResponse RespondToMouseDoubleClick(GH_Canvas sender, GH_CanvasMouseEvent e)
-      {
-        bool ctrl  = Control.ModifierKeys == Keys.Control;
-        bool shift = Control.ModifierKeys == Keys.Shift;
-
-        if (e.Button == MouseButtons.Left && (ctrl || shift))
-        {
-          if (!Owner.Params.Input[0].VolatileData.IsEmpty)
-          {
-            if (Owner is ElementDecompose elementDecompose)
-            {
-              sender.ActiveInteraction = null;
-              if (shift)
-                elementDecompose.Menu_PopulateOutputsWithCommonParameters(sender, e);
-              else if (ctrl)
-                elementDecompose.Menu_RemoveUnconnectedParameters(sender, e);
-
-              return GH_ObjectResponse.Handled;
-            }
-          }
-        }
-
-        return GH_ObjectResponse.Ignore;
-      }
-    }
-
-    public override void CreateAttributes() => m_attributes = new ComponentAttributes(this);
-
-    void AddOutputParameter(IGH_Param param)
-    {
-      if (param.Attributes is null)
-        param.Attributes = new GH_LinkedParamAttributes(param, Attributes);
-
-      param.Access = GH_ParamAccess.item;
-      Params.RegisterOutputParam(param);
-    }
-
-    void Menu_PopulateOutputsWithCommonParameters(object sender, EventArgs e)
-    {
-      var common = default(HashSet<Parameters.ParameterParam>);
-      foreach (var goo in Params.Input[0].VolatileData.AllData(true).OfType<Types.Element>())
-      {
-        var element = (DB.Element) goo;
-        if (element is null)
-          continue;
-
-        var current = new HashSet<Parameters.ParameterParam>();
-        foreach (var param in element.Parameters.OfType<DB.Parameter>())
-          current.Add(new Parameters.ParameterParam(param));
-
-        if (common is null)
-          common = current;
-        else
-          common.IntersectWith(current);
-      }
-
-      RecordUndoEvent("Get Common Parameters");
-
-      PopulateOutputParameters(common);
-    }
-
-    void Menu_PopulateOutputsWithAllParameters(object sender, EventArgs e)
-    {
-      var all = new HashSet<Parameters.ParameterParam>();
-      foreach (var goo in Params.Input[0].VolatileData.AllData(true).OfType<Types.Element>())
-      {
-        var element = (DB.Element) goo;
-        if (element is null)
-          continue;
-
-        foreach (var param in element.Parameters.OfType<DB.Parameter>())
-          all.Add(new Parameters.ParameterParam(param));
-      }
-
-      RecordUndoEvent("Get All Parameters");
-
-      PopulateOutputParameters(all);
-    }
-
-    void PopulateOutputParameters(IEnumerable<Parameters.ParameterParam> parameters)
-    {
-      var connected = new Dictionary<Parameters.ParameterParam, IList<IGH_Param>>();
-      foreach (var output in Params.Output.ToArray())
-      {
-        if
-        (
-          output.Recipients.Count > 0 &&
-          output is Parameters.ParameterParam param
-        )
-          connected.Add(param, param.Recipients.ToArray());
-
-        Params.UnregisterOutputParameter(output);
-      }
-
-      if (parameters is object)
-      {
-        foreach (var group in parameters.GroupBy(x => x.ParameterGroup).OrderBy(x => x.Key))
-        {
-          foreach (var parameter in group.OrderBy(x => x.ParameterBuiltInId))
-          {
-            AddOutputParameter(parameter);
-
-            if (connected.TryGetValue(parameter, out var recipients))
-            {
-              foreach (var recipient in recipients)
-                recipient.AddSource(parameter);
-            }
-          }
-        }
-      }
-
-      Params.OnParametersChanged();
-      ExpireSolution(true);
-    }
-
-    void Menu_RemoveUnconnectedParameters(object sender, EventArgs e)
-    {
-      RecordUndoEvent("Remove Unconnected Outputs");
-
-      foreach (var output in Params.Output.ToArray())
-      {
-        if (output.Recipients.Count > 0)
-          continue;
-
-        Params.UnregisterOutputParameter(output);
-      }
-
-      Params.OnParametersChanged();
-      OnDisplayExpired(false);
-    }
-
-    protected override void RegisterInputParams(GH_InputParamManager manager)
-    {
-      manager.AddParameter(new Parameters.Element(), "Element", "E", "Element to inspect", GH_ParamAccess.item);
-    }
-
-    protected override void RegisterOutputParams(GH_OutputParamManager manager) { }
-
-    protected override void TrySolveInstance(IGH_DataAccess DA)
-    {
-      DB.Element element = null;
-      if (!DA.GetData("Element", ref element))
-        return;
-
-      for (int p = 0; p < Params.Output.Count; ++p)
-      {
-        if (Params.Output[p] is Parameters.ParameterParam param)
-          DA.SetData(p, param.GetParameter(element));
-      }
-    }
-
-    bool IGH_VariableParameterComponent.CanInsertParameter(GH_ParameterSide side, int index) => false;
-    bool IGH_VariableParameterComponent.CanRemoveParameter(GH_ParameterSide side, int index) => side == GH_ParameterSide.Output;
-    IGH_Param IGH_VariableParameterComponent.CreateParameter(GH_ParameterSide side, int index) => null;
-    bool IGH_VariableParameterComponent.DestroyParameter(GH_ParameterSide side, int index) => true;
-    void IGH_VariableParameterComponent.VariableParameterMaintenance() { }
-
-    protected override string HtmlHelp_Source()
-    {
-      var nTopic = new Grasshopper.GUI.HTML.GH_HtmlFormatter(this)
-      {
-        Title = Name,
-        Description =
-        @"<p>This component is a special interface object that allows for quick accessing to Revit Element parameters.</p>" +
-        @"<p>It's able to modify itself in order to show any parameter its input element parameter contains. " +
-        @"It also allows to remove some output parameters if are not connected to anything else.</p>" +
-        @"<p>Under the component contextual menu you would find these options:</p>" +
-        @"<dl>" +
-        @"<dt><b>Get common parameters</b></dt><dd>Populates the output parameters list with common parameters in all input elements</dd>" +
-        @"<dt><b>Get all parameters</b></dt><dd>Populates the output parameters list with all parameters found in all input elements</dd>" +
-        @"<dt><b>Remove unconnected parameters</b></dt><dd>Removes the output parameters that are not connected to anything else</dd>" +
-        @"</dl>",
-        ContactURI = @"https://www.rhino3d.com/inside/revit/beta/"
-      };
-
-      nTopic.AddRemark("SHIFT + Double click runs \"Get common parameters\"");
-      nTopic.AddRemark("CTRL + Double click runs \"Remove unconnected parameters\".");
-
-      return nTopic.HtmlFormat();
     }
   }
 }
